moved styled components to the separate files
