+++ src/components/Collections/Cluster/StyledComponents/StyledNode.jsx
@@ -1,0 +1,8 @@
+import { styled } from '@mui/material/styles';
+
+export const StyledNode = styled('div')(() => ({
+  display: 'flex',
+  flexDirection: 'column',
+  gap: '0.5rem',
+}));
+
